# packages/api-contracts/openapi/gate0.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Thai Meet Gate 0 API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/chats/rooms/{roomId}": {
      "get": {
        "operationId": "getChatRoom",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat room with participant snapshot and messages"
          }
        }
      }
    },
    "/api/v1/chats/rooms/{roomId}/contact-exchanges/line": {
      "post": {
        "operationId": "createLineContactExchange",
        "x-error-codes": [
          "TM_API_CONTACT_EXCHANGE_MISSING_LINE_CONTACT",
          "TM_API_CONTACT_EXCHANGE_ROOM_MEMBERSHIP_REQUIRED",
          "TM_API_CONTACT_EXCHANGE_DUPLICATE_IDEMPOTENCY_KEY",
          "TM_API_CONTACT_EXCHANGE_REVOKED",
          "TM_API_CONTACT_EXCHANGE_REPORTED",
          "TM_API_CONTACT_EXCHANGE_BLOCKED",
          "TM_API_CONTACT_PROVIDER_UNAVAILABLE"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "locked",
                "available",
                "revoked",
                "reported",
                "blocked",
                "provider_unavailable"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "LINE ContactExchange created without exposing raw contact values"
          },
          "400": {
            "description": "LINE contact is missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Room membership is required or ContactExchange is blocked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate, revoked, or reported ContactExchange",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Contact provider is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "type",
              "code",
              "message",
              "param",
              "requestId",
              "docRef"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "validation_error",
                  "auth_error",
                  "permission_error",
                  "conflict_error",
                  "not_found",
                  "provider_error",
                  "system_error"
                ]
              },
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "param": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "docRef": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}